+++ packages/components/components/elvis-radio-filter/src/react/styledComponents.tsx
@@ -3,7 +3,7 @@ import ElviaColors from '@elvia/elvis-colors';
 
 const colors = {
   elviaCharge: ElviaColors['primary-colors']['green'].color,
-  elviaOff: ElviaColors['primary-colors']['black'].color,
+  elviaBlack: ElviaColors['primary-colors']['black'].color,
   outline: ElviaColors['internal-colors']['focus-outline'].color,
 };
 
@@ -18,12 +18,12 @@ export const RadioFilterLabel = styled.label`
   border-radius: 32px;
   padding: 4px 12px;
   border: ${(props: { isSelected: boolean }) =>
-    `1px solid ${props.isSelected ? colors.elviaOff : 'transparent'}`};
+    `1px solid ${props.isSelected ? colors.elviaBlack : 'transparent'}`};
   cursor: pointer;
 
   &:hover:not(:disabled) {
     border: ${(props: { isSelected: boolean }) =>
-      `1px solid ${props.isSelected ? colors.elviaOff : colors.elviaCharge}`};
+      `1px solid ${props.isSelected ? colors.elviaBlack : colors.elviaCharge}`};
   }
   &:focus-within {
     outline: 2px solid ${colors.outline};